backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
boto3==1.28.57
gunicorn==21.2.0
pydantic==2.11.3
pydantic-settings==2.9.1 